@obsoletesoftware-composition-analysis/python/python-example/requirements.txt
google-cloud-bigquery
python-dotenv
certifi
psutil